requirements.txt
torch>=0.4.0
torchvision>=0.2.1
dominate>=2.3.1
visdom>=0.1.8.3
scipy>=1.1.0
numpy>=1.14.1
Pillow>=5.0.0
opencv-python>=3.4.2